requirements.txt
Pillow==6.2.1
pkg-resources==0.0.0
textwrap3==0.9.2
numpy==1.14.5
